Type --> GeoJsonType
